Separate shared signing key infrastructure
