+++ src/lib/webbotauth.ts
@@ -22,12 +22,10 @@
 import {
   directoryResponseHeaders,
   HTTP_MESSAGE_SIGNATURES_DIRECTORY,
-  jwkToKeyID,
   MediaType,
-  type Signer,
   signatureHeaders,
 } from "web-bot-auth"
-import { helpers, signerFromJWK } from "web-bot-auth/crypto"
+import { configureSigningKey, currentSigningKey, type PublicSigningKey } from "./auth"
 import { PUBLISHER_ORIGIN } from "./identity"
 
 export interface WebBotAuthEnv {
@@ -57,28 +55,7 @@ const OUTBOUND_SIGNATURE_TTL_MS = 5 * 60 * 1000
 /** How long the directory self-signature stays valid. */
 const DIRECTORY_SIGNATURE_TTL_MS = 60 * 60 * 1000
 
-/** A public JSON Web Key as published in the directory (RFC 8037 Ed25519). */
-export interface DirectoryKey {
-  kty: "OKP"
-  crv: "Ed25519"
-  x: string
-  kid: string
-  use: string
-}
-
-interface WebBotAuthConfig {
-  signer: Signer
-  publicKey: DirectoryKey
-  agent: string
-}
-
-/**
- * Cached, parsed configuration.
- * The signing key is application-global (the same for every request),
- * so it is memoized across requests in the isolate
- * and only rebuilt when the underlying secret or agent changes.
- */
-let cache: { key: string; agent: string; config: Promise<WebBotAuthConfig> } | null = null
+let signatureAgent: string | null = null
 
 /**
  * Prime the signing configuration from the environment.
@@ -87,97 +64,8 @@ let cache: { key: string; agent: string; config: Promise<WebBotAuthConfig> } | n
  */
 export function configureWebBotAuth(env: WebBotAuthEnv): void {
   const key = env.WEB_BOT_AUTH_KEY?.trim()
-  if (!key) {
-    cache = null
-    return
-  }
-
-  const agent = env.SIGNATURE_AGENT?.trim() || DEFAULT_SIGNATURE_AGENT
-  if (cache && cache.key === key && cache.agent === agent) {
-    return
-  }
-
-  cache = { key, agent, config: buildConfig(key, agent) }
-}
-
-interface Ed25519PrivateJwk {
-  kty: "OKP"
-  crv: "Ed25519"
-  x: string
-  d: string
-}
-
-/**
- * Parse and validate the signing secret as an Ed25519 OKP private JWK.
- * Fails closed on anything malformed
- * so a misconfigured `WEB_BOT_AUTH_KEY` never yields a broken directory key
- * or unverifiable signatures.
- */
-function parseSigningJwk(key: string): Ed25519PrivateJwk {
-  let jwk: JsonWebKey
-  try {
-    jwk = JSON.parse(key) as JsonWebKey
-  } catch (error) {
-    throw new Error(`WEB_BOT_AUTH_KEY is not valid JSON: ${(error as Error).message}`)
-  }
-
-  if (
-    jwk.kty !== "OKP" ||
-    jwk.crv !== "Ed25519" ||
-    typeof jwk.x !== "string" ||
-    jwk.x.length === 0 ||
-    typeof jwk.d !== "string" ||
-    jwk.d.length === 0
-  ) {
-    throw new Error(
-      "WEB_BOT_AUTH_KEY must be an Ed25519 private JSON Web Key (kty=OKP, crv=Ed25519, with x and d).",
-    )
-  }
-
-  return { kty: jwk.kty, crv: jwk.crv, x: jwk.x, d: jwk.d }
-}
-
-async function buildConfig(key: string, agent: string): Promise<WebBotAuthConfig> {
-  const jwk = parseSigningJwk(key)
-  const signer = await signerFromJWK(jwk)
-  const kid = await jwkToKeyID(jwk, helpers.WEBCRYPTO_SHA256, helpers.BASE64URL_DECODE)
-
-  // Publish only the public half of the key (never the `d` parameter).
-  const publicKey: DirectoryKey = {
-    kty: jwk.kty,
-    crv: jwk.crv,
-    x: jwk.x,
-    kid,
-    use: "sig",
-  }
-
-  return { signer, publicKey, agent }
-}
-
-async function currentConfig(): Promise<WebBotAuthConfig | null> {
-  const entry = cache
-  if (!entry) {
-    return null
-  }
-
-  try {
-    return await entry.config
-  } catch (error) {
-    console.error("web-bot-auth: failed to load signing key", error)
-    // Drop the failed config so the next request rebuilds it,
-    // rather than serving the cached rejection until the isolate restarts.
-    // Guard against clobbering a newer config set by a concurrent reconfigure.
-    if (cache === entry) {
-      cache = null
-    }
-    return null
-  }
-}
-
-/** Return a copy of the configured public signing key, if available. */
-export async function webBotAuthPublicKey(): Promise<DirectoryKey | null> {
-  const config = await currentConfig()
-  return config ? { ...config.publicKey } : null
+  configureSigningKey(key)
+  signatureAgent = key ? env.SIGNATURE_AGENT?.trim() || DEFAULT_SIGNATURE_AGENT : null
 }
 
 /** Serialize a value as an RFC 8941 structured-field string (a quoted string). */
@@ -186,7 +74,7 @@ function structuredString(value: string): string {
 }
 
 export interface DirectoryResponse {
-  body: { keys: DirectoryKey[] }
+  body: { keys: PublicSigningKey[] }
   headers: Record<string, string>
 }
 
@@ -200,8 +88,8 @@ export interface DirectoryResponse {
  * demonstrating control of the published key.
  */
 export async function webBotAuthDirectory(requestUrl: string): Promise<DirectoryResponse | null> {
-  const config = await currentConfig()
-  if (!config) {
+  const signingKey = await currentSigningKey()
+  if (!signingKey) {
     return null
   }
 
@@ -212,10 +100,13 @@ export async function webBotAuthDirectory(requestUrl: string): Promise<Directory
     request: { method: "GET", url: requestUrl, headers: {} as Record<string, string> },
   }
 
-  const signature = await directoryResponseHeaders(message, [config.signer], { created, expires })
+  const signature = await directoryResponseHeaders(message, [signingKey.signer], {
+    created,
+    expires,
+  })
 
   return {
-    body: { keys: [config.publicKey] },
+    body: { keys: [signingKey.publicKey] },
     headers: {
       "Content-Type": DIRECTORY_MEDIA_TYPE,
       Signature: signature.Signature,
@@ -235,13 +126,14 @@ export async function webBotAuthHeaders(
   method: string,
   url: string | URL,
 ): Promise<Record<string, string>> {
-  const config = await currentConfig()
-  if (!config) {
+  const signingKey = await currentSigningKey()
+  const agentValue = signatureAgent
+  if (!signingKey || !agentValue) {
     return {}
   }
 
   try {
-    const agent = structuredString(config.agent)
+    const agent = structuredString(agentValue)
     // `signature-agent` must be present on the message
     // so it is covered by the signature;
     // the same value is returned for the outbound request.
@@ -252,7 +144,7 @@ export async function webBotAuthHeaders(
     const expires = new Date(created.getTime() + OUTBOUND_SIGNATURE_TTL_MS)
     const signature = await signatureHeaders(
       { method, url: url.toString(), headers },
-      config.signer,
+      signingKey.signer,
       { created, expires },
     )
